requirements.txt
PyQt6>=6.6,<7.0
Pillow>=10.0,<11.0
